% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_hs_player_details.R
\name{get_all_hs_player_details}
\alias{get_all_hs_player_details}
\title{Parse Tracking Football Data for High School Players}
\usage{
get_all_hs_player_details(
  api_key,
  base_url = "https://tfb2-api.trackingfootball.com/api/v1/api/",
  page_size = 500,
  verbose = TRUE,
  sleep_sec = 0.1,
  updated_since = "2020-01-01"
)
}
\arguments{
\item{api_key}{A character element containing an API key obtained from Tracking Football}

\item{base_url}{Tracking Football APi base URL: "https://tfb2-api.trackingfootball.com/api/v1/api/"}

\item{page_size}{Number of players per page}

\item{verbose}{Print Progress messages}

\item{sleep_sec}{Sleep Timer for API call}

\item{updated_since}{Date parameter for oldest date to query API in format "YYYY-MM-DD"}
}
\value{
all player details as a list of JSON-decoded R objects
}
\description{
Parse Tracking Football Data for High School Players
}
