% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_awards.R
\name{get_awards}
\alias{get_awards}
\title{Parse Award List data from Tracking Football Player List}
\usage{
get_awards(player_list)
}
\arguments{
\item{player_list}{all player details list JSON-decoded object obtained through one of the get_all_xyz_player_details functions}
}
\value{
all players listed on College or NFL awards lists
}
\description{
Parse Award List data from Tracking Football Player List
}
\examples{
\dontrun{
df <- get_awards(player_list)
}
}
