% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bestTrackEvent.R
\name{get_bestTrackEvent}
\alias{get_bestTrackEvent}
\title{Parse bestTrackEvent data from Tracking Football Player List}
\usage{
get_bestTrackEvent(player_list)
}
\arguments{
\item{player_list}{all player details list JSON-decoded object obtained through one of the get_all_xyz_player_details functions}
}
\value{
all player bestTrackEvent details as a list object
}
\description{
Parse bestTrackEvent data from Tracking Football Player List
}
