% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collegeProduction.R
\name{get_collegeProduction}
\alias{get_collegeProduction}
\title{Parse College Production data from Tracking Football Player List}
\usage{
get_collegeProduction(player_list)
}
\arguments{
\item{player_list}{all player details list JSON-decoded object obtained through one of the get_all_xyz_player_details functions}
}
\value{
all player collegeProduction details as a list object
}
\description{
Parse College Production data from Tracking Football Player List
}
\examples{
\dontrun{
df <- get_collegeProduction(player_list)
}
}
