% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_highSchools.R
\name{get_highSchools}
\alias{get_highSchools}
\title{Parse highSchools data from Tracking Football Player List}
\usage{
get_highSchools(player_list)
}
\arguments{
\item{player_list}{all player details list JSON-decoded object obtained through one of the get_all_xyz_player_details functions}
}
\value{
all player highSchools details as a list object
}
\description{
Parse highSchools data from Tracking Football Player List
}
