% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hsCombines.R
\name{get_hsCombines}
\alias{get_hsCombines}
\title{Parse hsCombines data from Tracking Football Player List}
\usage{
get_hsCombines(player_list)
}
\arguments{
\item{player_list}{all player details list JSON-decoded object obtained through one of the get_all_xyz_player_details functions}
}
\value{
all player hsCombines details as a list object
}
\description{
Parse hsCombines data from Tracking Football Player List
}
