% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_links}
\alias{get_links}
\title{Parse links data from Tracking Football Player List}
\usage{
get_links(player_list)
}
\arguments{
\item{player_list}{all player details list JSON-decoded object obtained through one of the get_all_xyz_player_details functions}
}
\value{
all player links details as a list object
}
\description{
Parse links data from Tracking Football Player List
}
