% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multiSport.R
\name{get_multiSport}
\alias{get_multiSport}
\title{Parse Multi Sport data from Tracking Football Player List}
\usage{
get_multiSport(player_list)
}
\arguments{
\item{player_list}{all player details list JSON-decoded object obtained through one of the get_all_xyz_player_details functions}
}
\value{
High school sports played by all players listed
}
\description{
Parse Multi Sport data from Tracking Football Player List
}
\examples{
\dontrun{
df <- get_multiSport(player_list)
}
}
