% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_list.R
\name{get_player_list}
\alias{get_player_list}
\alias{get_all_player_details}
\title{Get Player List}
\usage{
get_player_list(
  api_key,
  base_url = "https://tfb2-api.trackingfootball.com/api/v1/api/",
  page_size = 500,
  verbose = TRUE,
  sleep_sec = 0.005,
  updated_since = as.character(Sys.Date() - 1),
  player_type = "college"
)
}
\arguments{
\item{api_key}{A character element containing an API key obtained from Tracking Football}

\item{base_url}{Tracking Football APi base URL: "https://tfb2-api.trackingfootball.com/api/v1/api/"}

\item{page_size}{Number of players per page}

\item{verbose}{Print Progress messages}

\item{sleep_sec}{Sleep Timer for API call}

\item{updated_since}{Date parameter for oldest date to query API in format "YYYY-MM-DD"}

\item{player_type}{Character, call in type of player. Options: "high-school", "college", "nfl". Default to "college"}
}
\value{
A large list with nested dataframes for individual player measurables and performances.
}
\description{
Retrieve Tracking Football player details in a structured list since a given date.
}
\examples{
\dontrun{
ncaa_historical <- get_player_list(api_key, updated_since = "2020-01-01", player_type = "college")
}
}
