% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_playerDetails.R
\name{get_playerDetails}
\alias{get_playerDetails}
\title{Parse playerDetails data from Tracking Football Player List}
\usage{
get_playerDetails(player_list)
}
\arguments{
\item{player_list}{all player details list JSON-decoded object obtained through one of the get_all_xyz_player_details functions}
}
\value{
all player playerDetails details as a list object
}
\description{
Parse playerDetails data from Tracking Football Player List
}
