% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_watchList.R
\name{get_watchList}
\alias{get_watchList}
\title{Parse College Watch List data from Tracking Football Player List}
\usage{
get_watchList(player_list)
}
\arguments{
\item{player_list}{all player details list JSON-decoded object obtained through one of the get_all_xyz_player_details functions}
}
\value{
all players listed on College all-star game watch lists
}
\description{
Parse College Watch List data from Tracking Football Player List
}
\examples{
\dontrun{
df <- get_watchList(player_list)
}
}
